+++ Source/ACE.Server/WorldObjects/Player_Xp.cs
@@ -407,6 +407,8 @@ namespace ACE.Server.WorldObjects
             var newItemLevel = item.ItemLevel.Value;
             if (newItemLevel > prevItemLevel)
             {
+                OnItemLevelUp(item, prevItemLevel);
+
                 var actionChain = new ActionChain();
                 actionChain.AddAction(this, () =>
                 {
@@ -414,9 +416,6 @@ namespace ACE.Server.WorldObjects
                     Session.Network.EnqueueSend(new GameMessageSystemChat(msg, ChatMessageType.Broadcast));
 
                     EnqueueBroadcast(new GameMessageScript(Guid, ACE.Entity.Enum.PlayScript.AetheriaLevelUp));
-
-                    OnItemLevelUp(item, prevItemLevel);
-
                 });
                 actionChain.EnqueueChain();
             }